Struct const_hex::Buffer

#[repr(C)]
pub struct Buffer<const N: usize, const PREFIX: bool = false> { /* private fields */ }

A correctly sized stack allocation for the formatted bytes to be written into.

N is the amount of bytes of the input, while PREFIX specifies whether the “0x” prefix is prepended to the output.

Note that this buffer will contain only the prefix, if specified, and null (‘\0’) bytes before any formatting is done.

Examples

let mut buffer = const_hex::Buffer::<4>::new();
let printed = buffer.format(b"1234");
assert_eq!(printed, "31323334");

Implementations

impl<const N: usize, const PREFIX: bool> Buffer<N, PREFIX>

pub const LEN: usize = _

The length of the buffer in bytes.

pub const fn new() -> Self

This is a cheap operation; you don’t need to worry about reusing buffers for efficiency.

pub const fn const_format(self, array: &[u8; N]) -> Self

Print an array of bytes into this buffer.

pub const fn const_format_upper(self, array: &[u8; N]) -> Self

Print an array of bytes into this buffer.

pub fn format(&mut self, array: &[u8; N]) -> &mut str

Print an array of bytes into this buffer and return a reference to its lower hex string representation within the buffer.

pub fn format_upper(&mut self, array: &[u8; N]) -> &mut str

Print an array of bytes into this buffer and return a reference to its upper hex string representation within the buffer.

pub fn format_slice<T: AsRef<[u8]>>(&mut self, slice: T) -> &mut str

Print a slice of bytes into this buffer and return a reference to its lower hex string representation within the buffer.

Panics

If the slice is not exactly N bytes long.

pub fn format_slice_upper<T: AsRef<[u8]>>(&mut self, slice: T) -> &mut str

Print a slice of bytes into this buffer and return a reference to its upper hex string representation within the buffer.

Panics

If the slice is not exactly N bytes long.

pub fn to_string(&self) -> String

Copies self into a new owned String.

pub const fn as_str(&self) -> &str

Returns a reference to the underlying bytes casted to a string slice.

pub fn as_mut_str(&mut self) -> &mut str

Returns a mutable reference to the underlying bytes casted to a string slice.

pub fn to_vec(&self) -> Vec<u8>

Copies self into a new Vec.

pub fn as_byte_array<const LEN: usize>(&self) -> &[u8; LEN]

Returns a reference the underlying stack-allocated byte array.

Panics

If LEN does not equal Self::LEN.

This is panic is evaluated at compile-time if the nightly feature is enabled, as inline const blocks are currently unstable.

See Rust tracking issue #76001.

pub fn as_mut_byte_array<const LEN: usize>(&mut self) -> &mut [u8; LEN]

Returns a mutable reference the underlying stack-allocated byte array.

Panics

If LEN does not equal Self::LEN.

See as_byte_array for more information.

pub const fn as_bytes(&self) -> &[u8]

Returns a reference to the underlying bytes.

pub unsafe fn as_mut_bytes(&mut self) -> &mut [u8]

Returns a mutable reference to the underlying bytes.

Safety

The caller must ensure that the content of the slice is valid UTF-8 before the borrow ends and the underlying str is used.

Use of a str whose contents are not valid UTF-8 is undefined behavior.

pub unsafe fn buffer(&mut self) -> &mut [u8]

Returns a mutable reference to the underlying buffer, excluding the prefix.

Safety

See as_mut_bytes.

pub const fn as_ptr(&self) -> *const u8

Returns a raw pointer to the buffer.

The caller must ensure that the buffer outlives the pointer this function returns, or else it will end up pointing to garbage.

pub fn as_mut_ptr(&mut self) -> *mut u8

Returns an unsafe mutable pointer to the slice’s buffer.

The caller must ensure that the slice outlives the pointer this function returns, or else it will end up pointing to garbage.